Refactoring du système d'items : - les items dans un namspace - tous les attributs en private - ajout du itemType
