--- Assets/Scripts/ItemInstances.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 8fb54beb5d0360a45a202a169a2aa184
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
