--- Assets/Scripts/ItemInstances/Hoe.asset
@@ -1,21 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5774d819f65697647a7d771ac551045e, type: 3}
-  m_Name: Hoe
-  m_EditorClassIdentifier: 
-  nameItem: Hoe
-  description: 
-  type: 0
-  visuel: {fileID: 21300000, guid: 4632e854ef0ee934885e0953c49553e6, type: 3}
-  prefab: {fileID: 1670586203085811935, guid: ae78e34dbecdca847a938199eb2badb3, type: 3}
-  stackable: 0
-  weight: 1000
