--- Assets/Scripts/ItemInstances/Hoe.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3cf3e31873b6fe042b076bd871c96b1f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
