--- Assets/Scripts/ItemInstances/Tomato.asset
@@ -1,21 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5774d819f65697647a7d771ac551045e, type: 3}
-  m_Name: Tomato
-  m_EditorClassIdentifier: 
-  nameItem: Tomato
-  description: 
-  type: 1
-  visuel: {fileID: 21300000, guid: 8ade532de32d9fd42ae7accbe310c943, type: 3}
-  prefab: {fileID: 3333028828918135912, guid: 387244667f3e96e4dbed87522ac27e49, type: 3}
-  stackable: 1
-  weight: 500
