--- Assets/Scripts/ItemInstances/Tomato.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 19d9001a6268ade44b6892098b3fd939
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
