+++ Assets/Scripts/ItemSystem.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 52a963017af543d3bb62b0d767306efe
+timeCreated: 1733833780
