+++ Assets/Scripts/ItemSystem/ItemData.cs.meta
