+++ Assets/Scripts/ItemSystem/ItemType.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: ebdf0b84fb2a43a4b3cdf225c2b66e21
+timeCreated: 1733833825
